+++ spec/dummy/app/controllers/application_controller.rb
@@ -53,6 +53,8 @@ class ApplicationController < ActionController::Base
 
   def editor_preview?; end
   helper_method :editor_preview?
-end
 
-def service_slug_config; end
+  def service_slug_config
+    service.service_slug
+  end
+end